Link in the database.yml file
We need to use the database.yml file that is in the shared directory,
not the one that has been checked into the repo.  Ansible is placing a
copy of the database.yml file into the shared directory - use this file
since it has the necessary credentials to access staging and production
databases.