Rename database.yml to database.example.yml
The database.yml file needs to be updated as we are no longer using ENV
variables to store the db credentials since rake had problems accessing
the values for the ENV variables when running tasks such as `db:seed`.
Since we are moving to a model where the credentials are stored in the
database.yml file, want to make sure we don't make it easy to commit
that file to the repository, so renaming it to database.example.yml and
telling git to ignore any file ending in database.yml.